feat(client): 태그 필터링 기능 추가
