+++ apps/client/components/home/index.tsx
@@ -16,8 +16,19 @@ import TagList from './tag-list';
 import { TReturnComponent } from '@/types/common.type';
 import { articleKeys } from '@/config/react-query/query-key-factory';
 import { useCallback } from 'react';
+import useCreateQueryString from '@/features/hooks/use-create-query-string';
 
 export default function HomeComponent(): TReturnComponent {
+    const { moveToUrlWithoutAnchor } = useCreateQueryString();
+
+    const searchParams = useSearchParams();
+
+    const currentPage = searchParams.get('page') || 1;
+    const showPageCnt = searchParams.get('limit') || 10;
+    const tag = searchParams.get('tag');
+    const author = searchParams.get('author');
+    const favorited = searchParams.get('favorited');
+
     const { data } = useQuery({
         queryKey: articleKeys.lists(),
         queryFn: async () =>
@@ -27,29 +38,11 @@ export default function HomeComponent(): TReturnComponent {
             ),
     });
 
-    const router = useRouter();
-    const searchParams = useSearchParams();
-    const pathname = usePathname();
-
-    const currentPage = searchParams.get('page') || 1;
-    const showPageCnt = searchParams.get('limit') || 10;
     const articleListData = data?.getArticleList.articles || [];
     const totalCnt = data?.getArticleList.articlesCount || 0;
 
-    // TODO: 여러 컴포넌트에서 반복 사용할 함수가 될 예정이므로 분리 필요
-    const createQueryString = useCallback(
-        (name: string, value: string) => {
-            const params = new URLSearchParams(searchParams);
-            params.set(name, value);
-
-            return params.toString();
-        },
-        [searchParams],
-    );
-
     const onClickPagination = (page: number) => {
-        const queryString = createQueryString('page', String(page));
-        router.push(pathname + '?' + queryString);
+        moveToUrlWithoutAnchor({ key: 'page', value: String(page) });
     };
 
     return (
